requirements.txt
# Core dependencies
torch>=2.0.0
transformers>=4.30.0
Pillow>=9.0.0
numpy>=1.21.0

# OCR and document processing
pytesseract>=0.3.10
PyMuPDF>=1.23.0
pdf2image>=1.17.0

# Image processing
opencv-python>=4.8.0
scikit-image>=0.21.0

# Machine learning
scikit-learn>=1.3.0

# Utilities
requests>=2.31.0
tqdm>=4.65.0
pathlib2>=2.3.7

# Development dependencies (optional)
# pytest>=6.0
# black>=21.0
# flake8>=3.8
# mypy>=0.800

# Optional: GPU support
# torchvision>=0.15.0

# Optional: Additional OCR engines
# olmocr[gpu]  # Uncomment if using OlmOCR
# easyocr  # Uncomment if using EasyOCR
